Clean up CalibrateEmulateSample
