--- examples/L96m/cfg.ini
@@ -1,16 +1,0 @@
-[Runs]
-run_conv = true
-run_dns  = true
-run_bal  = true
-run_reg  = true
-
-[Integration]
-SOLVER = Tsit5
-
-T      = 4.0
-T_conv = 100.0
-
-dtmax  = 1e-3
-reltol = 1e-3
-abstol = 1e-6
-
