--- examples/Project.toml
@@ -1,9 +1,0 @@
-[deps]
-DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
-Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-IJulia = "7073ff75-c697-5162-941a-fcdaad2a7d2a"
-LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
-PyPlot = "d330b81b-6aea-500a-939a-2ce795aea3ee"
-CalibrateEmulateSample = "16b9c28e-565c-11e9-13a9-53e2645fa95d"
