+++ src/CalibrateEmulateSample.jl
@@ -4,14 +4,10 @@ using Distributions, Statistics, LinearAlgebra, DocStringExtensions
 
 # No internal deps, light external deps
 include("Observations.jl")
-include("spaces.jl")
-include("problems.jl")
-include("EKS.jl")
 include("Priors.jl")
 
 # No internal deps, heavy external deps
 include("EKI.jl")
-include("GModel.jl")
 include("GPEmulator.jl")
 
 # Internal deps, light external deps
@@ -19,8 +15,5 @@ include("Utilities.jl")
 
 # Internal deps, light external deps
 include("MCMC.jl")
-# include("Histograms.jl")
-# include("GPR.jl")
-
 
 end # module
